+++ app/src/main/java/nimonscooked/entity/item/Item.java
@@ -4,7 +4,7 @@ import nimonscooked.entity.Entity;
 
 public abstract class Item extends Entity {
 
-    public Item(String id, String name, float x, float y) {
+    public Item(String id, String name, int x, int y) {
         super(id, name, x, y);
     }
