+++ app/src/main/java/nimonscooked/entity/item/ingredient/Tomato.java
@@ -3,7 +3,7 @@ package nimonscooked.entity.item.ingredient;
 import nimonscooked.enums.IngredientState;
 
 public class Tomato extends Ingredient{
-    public Tomato(String id, float x, float y, IngredientState currentState) {
+    public Tomato(String id, int x, int y, IngredientState currentState) {
         super(id, "Tomato", x, y, currentState);
     }
